add SMTP settings helper
